\name{ipc}
\alias{ipc}
\docType{data}
\title{Brazilian Price Consumer Index (IPC-DI)}
\description{
The Consumer Price Index (IPC-DI) measures the change in prices of a fixed set of goods and services components of usual expenses of families with income level of between 1 and 33 minimum monthly wages. It is published by the Brazilian Institute of Economics (FGV|IBRE).
}
\usage{
data(ipc)
}
\format{
A monthly time series, in variation (\%).
Period: jan/2004 - may/2015.
}
\source{
\url{http://portalibre.fgv.br/}
}
\examples{
data(ipc)
plot(ipc)
}

